NeuralGuard/requirements.txt
kafka-python==2.0.2
elasticsearch==7.17.9      
numpy==1.23.5      
requests==2.31.0          
scipy==1.10.1     
scikit-learn==1.3.0        #
